refactor: split index reference dtos
