+++ docs/standards/temporal-vocabulary-allowlist.json
@@ -42,9 +42,37 @@
         {
             "path": "src/services/ingestion_service/app/DTOs/reference_data_index_series_dto.py",
             "field": "source_timestamp",
-            "expectedCount": 4,
-            "ownerSlice": "Slice 7 / CR-1046",
-            "reason": "Existing index definition, index price, index return, and risk-free ingestion DTO fields moved from reference_data_benchmark_dto.py during index-series DTO modularization. New source-observation fields must use observed_at."
+            "expectedCount": 0,
+            "ownerSlice": "Slice 7 / CR-1046 / CR-1053",
+            "reason": "Legacy index definition, index price, index return, and risk-free ingestion DTO fields have been moved into focused DTO-family modules. New source-observation fields must use observed_at."
+        },
+        {
+            "path": "src/services/ingestion_service/app/DTOs/reference_data_index_definition_dto.py",
+            "field": "source_timestamp",
+            "expectedCount": 1,
+            "ownerSlice": "Slice 7 / CR-1053",
+            "reason": "Existing index definition ingestion DTO field moved from reference_data_index_series_dto.py during index/risk-free DTO modularization. New source-observation fields must use observed_at."
+        },
+        {
+            "path": "src/services/ingestion_service/app/DTOs/reference_data_index_price_series_dto.py",
+            "field": "source_timestamp",
+            "expectedCount": 1,
+            "ownerSlice": "Slice 7 / CR-1053",
+            "reason": "Existing index price-series ingestion DTO field moved from reference_data_index_series_dto.py during index/risk-free DTO modularization. New source-observation fields must use observed_at."
+        },
+        {
+            "path": "src/services/ingestion_service/app/DTOs/reference_data_index_return_series_dto.py",
+            "field": "source_timestamp",
+            "expectedCount": 1,
+            "ownerSlice": "Slice 7 / CR-1053",
+            "reason": "Existing index return-series ingestion DTO field moved from reference_data_index_series_dto.py during index/risk-free DTO modularization. New source-observation fields must use observed_at."
+        },
+        {
+            "path": "src/services/ingestion_service/app/DTOs/reference_data_risk_free_series_dto.py",
+            "field": "source_timestamp",
+            "expectedCount": 1,
+            "ownerSlice": "Slice 7 / CR-1053",
+            "reason": "Existing risk-free series ingestion DTO field moved from reference_data_index_series_dto.py during index/risk-free DTO modularization. New source-observation fields must use observed_at."
         },
         {
             "path": "src/services/ingestion_service/app/DTOs/reference_data_support_dto.py",
